requirements.txt
web3
python-telegram-bot
psycopg2-binary
python-dotenv
pytest
pytest-mock
schedule